CertificateService package name refactored
