+++ rest-api/web/src/main/resources/locator.xml
@@ -40,8 +40,8 @@
         <api>org.eclipse.kapua.service.authorization.role.RolePermissionFactory</api>
         <api>org.eclipse.kapua.service.authorization.role.RolePermissionService</api>
 
-        <api>org.eclipse.kapua.service.certificate.api.CertificateService</api>
-        <api>org.eclipse.kapua.service.certificate.api.CertificateFactory</api>
+        <api>org.eclipse.kapua.service.certificate.CertificateService</api>
+        <api>org.eclipse.kapua.service.certificate.CertificateFactory</api>
 
         <api>org.eclipse.kapua.service.datastore.model.StorableIdFactory</api>
