+++ service/security/certificate/api/src/main/java/org/eclipse/kapua/service/certificate/Certificate.java
@@ -9,12 +9,13 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.service.certificate.api;
+package org.eclipse.kapua.service.certificate;
 
-import org.eclipse.kapua.model.KapuaNamedEntityCreator;
+import org.eclipse.kapua.model.KapuaUpdatableEntity;
 import org.eclipse.kapua.model.xml.DateXmlAdapter;
-import org.eclipse.kapua.service.certificate.api.xml.PrivateKeyXmlAdapter;
-import org.eclipse.kapua.service.certificate.api.xml.X509CertificateXmlAdapter;
+import org.eclipse.kapua.service.certificate.xml.CertificateXmlRegistry;
+import org.eclipse.kapua.service.certificate.xml.PrivateKeyXmlAdapter;
+import org.eclipse.kapua.service.certificate.xml.X509CertificateXmlAdapter;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
@@ -25,12 +26,7 @@ import java.security.PrivateKey;
 import java.security.cert.X509Certificate;
 import java.util.Date;
 
-/**
- * CertificateCreator encapsulates all the information needed to create a new Certificate in the system.
- *
- * @since 1.0
- */
-@XmlRootElement(name = "userCreator")
+@XmlRootElement(name = "certificate")
 @XmlAccessorType(XmlAccessType.PROPERTY)
 @XmlType(propOrder = {
         "certificate",
@@ -49,8 +45,14 @@ import java.util.Date;
         "ca",
         "caId",
         "password"
-})
-public interface CertificateCreator extends KapuaNamedEntityCreator<Certificate> {
+}, factoryClass = CertificateXmlRegistry.class, factoryMethod = "newCertificate")
+public interface Certificate extends KapuaUpdatableEntity {
+
+    String TYPE = "account";
+
+    default String getType() {
+        return TYPE;
+    }
 
     @XmlJavaTypeAdapter(X509CertificateXmlAdapter.class)
     X509Certificate getCertificate();
@@ -64,7 +66,7 @@ public interface CertificateCreator extends KapuaNamedEntityCreator<Certificate>
 
     Integer getVersion();
 
-    void setVersion(Integer family);
+    void setVersion(Integer version);
 
     String getSerial();
 
@@ -121,4 +123,5 @@ public interface CertificateCreator extends KapuaNamedEntityCreator<Certificate>
     String getPassword();
 
     void setPassword(String password);
+
 }
