+++ service/security/certificate/api/src/main/java/org/eclipse/kapua/service/certificate/CertificateFactory.java
@@ -9,7 +9,7 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.service.certificate.api;
+package org.eclipse.kapua.service.certificate;
 
 import org.eclipse.kapua.model.KapuaEntityFactory;
