+++ service/security/certificate/api/src/main/java/org/eclipse/kapua/service/certificate/CertificateStatus.java
@@ -9,7 +9,7 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.service.certificate.api;
+package org.eclipse.kapua.service.certificate;
 
 public enum CertificateStatus {
     VALID,
